Sample refactoring as class per sub-sample
