+++ cs/samples/StoreVarLenTypes/AsciiSumSample.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license.
 
 using System;
-using System.Diagnostics;
 using System.Linq;
 using FASTER.core;
